checker: Add `GetClients` accessor
